Move libovsdbops package
Move the libovsdbops package out of the
`go-controller/pkg/ovn` directory and into
`go-controller/pkg` since its used by many packages
in the directory.
