+++ go-controller/pkg/libovsdbops/datapath.go
