+++ go-controller/pkg/libovsdbops/switch.go
