+++ go-controller/pkg/ovn/loadbalancer/lb_cache.go
@@ -6,7 +6,8 @@ import (
 	"sync"
 
 	libovsdbclient "github.com/ovn-org/libovsdb/client"
-	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/ovn/libovsdbops"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/libovsdbops"
+
 	"k8s.io/apimachinery/pkg/util/sets"
 )
