+++ go-controller/pkg/util/ovn.go
@@ -11,7 +11,8 @@ import (
 	libovsdbclient "github.com/ovn-org/libovsdb/client"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/config"
 
-	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/ovn/libovsdbops"
+	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/libovsdbops"
+
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/sbdb"
 )
